requirements.txt
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
openai>=1.30.0
python-dotenv>=1.0.0
pyyaml>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
jupyter>=1.0.0
